refactor: 대시보드 디렉터리 분리 -> 파일 이동
