+++ src/features/dashboard/components/PopularPosts.vue
