+++ src/features/dashboard/components/chart/AlgorithmChart.vue
