+++ src/features/dashboard/components/chart/AverageChart.vue
