+++ src/features/dashboard/components/chart/FollowerChart.vue
