+++ src/features/dashboard/components/chart/SummaryItem.vue
