+++ src/features/dashboard/views/CampaginResultDashboardView.vue
