+++ src/features/dashboard/views/YoutubeDashboardView.vue
@@ -3,46 +3,44 @@ import { ref, onMounted, computed } from 'vue'
 import { useRoute, useRouter } from 'vue-router';
 import { Icon } from '@iconify/vue'
 
-import DashboardBase from '@/features/influencer/components/DashboardBase.vue'
-import DashboardHeader from '@/features/influencer/components/DashboardHeader.vue'
-import PopularPosts from '@/features/influencer/components/PopularPosts.vue'
-import PopularShortForms from '@/features/influencer/components/PopularShortForms.vue'
-import DashboardCampaignList from '@/features/influencer/components/DashboardCampaignList.vue'
-import { useToast } from 'vue-toastification';
+import DashboardBase from '@/features/dashboard/components/DashboardBase.vue'
+import DashboardHeader from '@/features/dashboard/components/DashboardHeader.vue'
+import PopularPosts from '@/features/dashboard/components/PopularPosts.vue'
+import PopularShortForms from '@/features/dashboard/components/PopularShortForms.vue'
+import DashboardCampaignList from '@/features/dashboard/components/DashboardCampaignList.vue'
 
 const route = useRoute()
 const router = useRouter()
-const toast = useToast()
 const data = ref(null)
-const satisfaction = ref(92.5)
+const satisfaction = ref(82.5)
 
 const influencerId = route.query.id
 
 onMounted(async () => {
-  const res = await fetch(`/api/v1/dashboard/instagram?id=${influencerId}`)
+  const res = await fetch(`/api/v1/dashboard/youtube?id=${influencerId}`)
   const resData = await res.json()
 
   if (!resData.data) {
-    toast.warning('인스타그램 계정이 연결되어있지 않습니다.');
-    router.replace(`/influencer/dashboard/youtube?id=${influencerId}`);
+    toast.warning('유튜브 계정이 연결되어있지 않습니다.');
+    router.replace(`/influencer/dashboard/instagram?id=${influencerId}`);
   }
 
   data.value = resData.data
 })
 
 const summaryData = computed(() => {
-  if (!data.value?.reelsSummary) {
+  if (!data.value?.shortsSummary) {
     return { shorts: 0, views: '0만', comments: '0개', likes: '0만' }
   }
   return {
-    shorts: data.value.reelsSummary.count,
-    views: `${(data.value.reelsSummary.avgViews / 10000).toFixed(1)}만`,
-    comments: `${data.value.reelsSummary.avgComments}개`,
-    likes: `${(data.value.reelsSummary.avgLikes / 10000).toFixed(1)}만`
+    shorts: data.value.shortsSummary.count,
+    views: `${(data.value.shortsSummary.averageViewCount / 10000).toFixed(1)}만`,
+    comments: `${data.value.shortsSummary.commentCount}개`,
+    likes: `${(data.value.shortsSummary.likeCount / 10000).toFixed(1)}만`
   }
 })
 
-const formatFollowers = (num) => {
+const formatSubscribers = (num) => {
   const parsedNum = parseInt(num)
   return parsedNum >= 10000 ? `${Math.floor(parsedNum / 10000)}만명` : `${parsedNum}명`
 }
@@ -69,14 +67,14 @@ const goToList = () => {
         </button>
       </div>
       <DashboardHeader
-        :name="data.title"
-        :thumbnail="data.thumbnail"
+        :name="data.channel.title"
+        :thumbnail="data.channel.thumbnails.default.url"
         :tags="data.tags"
-        :instaFollowers="formatFollowers(data.instagram.followers)"
+        :subscribers="formatSubscribers(data.channel.statistics.subscriberCount)"
       />
 
       <DashboardBase
-        :platform="'instagram'"
+        :platform="'youtube'"
         :summaryData="summaryData"
         :data="data"
         :satisfaction="satisfaction"
@@ -84,13 +82,13 @@ const goToList = () => {
       />
 
       <PopularPosts
-        :platform="'instagram'"
-        :items="data.popularPosts"
+        :platform="'youtube'"
+        :items="data.popularVideos"
       />
 
       <PopularShortForms
-        :platform="'instagram'"
-        :items="data.popularReels"
+        :platform="'youtube'"
+        :items="data.popularShorts"
       ></PopularShortForms>
 
       <DashboardCampaignList />
